// SIROS.Web/ClientApp/src/app/models/empresa.model.ts
export class EmpresaModel {

    nIdEmpresa?: number = 0;
    sRuc: string;
    sRazonSocial: string;
    nIdTipoVia: number;
    sNombreVia: string;
    nIdCcpp: number;
    sNombreCcpp: string;
    nIdNroMzKm: number;
    sNroMzKm: string;
    nIdLoteIntDpto: number;
    sLoteIntDpto: string;
    sReferencia: string;
    sCodDistrito: string;
    sCodProvincia: string;
    sCodDepartamento: string;
    sTelefono: string;
    sRepresentante: string;
    sUsuario?: string = '';
    nIdSession?: number = 0;
    nIdEntidadSSO?: number = 0;

}